fix: convert NumPy float64 to Python float in AI model predictions
CRITICAL DATABASE FIX: PostgreSQL was rejecting inserts because the AI model
was returning np.float64 objects instead of Python floats.

Error: schema "np" does not exist - PostgreSQL thought np.float64(99.44)
was a schema reference instead of a number value.

Changes:
- Wrap all return values with explicit float() conversion
- confidence_normal: float(round(float(confidence_normal), 2))
- confidence_anomaly: float(round(float(confidence_anomaly), 2))
- anomaly_score: float(round(float(score), 4))

This ensures all numeric values are native Python floats that PostgreSQL
can properly serialize in SQL INSERT statements.

Impact:
- WebSocket data will now save to database successfully
- Frontend will receive real-time metrics
- AI predictions will be stored correctly

+++ backend/fastapi/api/ai_model/model.py
@@ -111,9 +111,9 @@ def predict(heart_rate: float, motion_intensity: float):
 
         return {
             "prediction": status,
-            "anomaly_score": round(float(score), 4),
-            "confidence_normal": round(confidence_normal, 2),
-            "confidence_anomaly": round(confidence_anomaly, 2)
+            "anomaly_score": float(round(float(score), 4)),
+            "confidence_normal": float(round(float(confidence_normal), 2)),
+            "confidence_anomaly": float(round(float(confidence_anomaly), 2))
         }
 
     except Exception as e:
